move helpers to utils.js
and fix duplicate declaration in addListing scenario
